IAM_TASK_3/assignment_files/requirements.txt
cv2
pandas
re
Pillow
numpy
json
string
sklearn
skimage
scipy
pickle
